refactor: change path for local storage
