+++ src/requests/auth.test.ts
@@ -1,28 +1,8 @@
 import { authenticateUser, requestToSignIn, requestToSignUp } from "./auth";
-import { userStorage } from "@/utils/storage/index";
+import { userStorage } from "@/storages";
 import CustomError from "@/utils/common/errors/CustomError";
 import { createClient } from "@/utils/supabase/component";
 import { UserResponse, AuthError } from "@supabase/supabase-js";
-
-jest.mock("@/utils/supabase/component", () => ({
-  createClient: jest.fn(() => ({
-    auth: {
-      getUser: jest.fn(),
-      signInWithPassword: jest.fn(),
-      signUp: jest.fn(),
-    },
-  })),
-}));
-
-jest.mock("@/utils/storage/index", () => ({
-  userStorage: {
-    set: jest.fn(),
-    get: jest.fn(() => ({
-      id: "user-id",
-      email: "user@example.com",
-    })),
-  },
-}));
 
 const USER_DUMMY = {
   id: "user-id",
@@ -39,17 +19,32 @@ const USER_DUMMY = {
     sub: "ba1acab4-e453-4b0d-9cb5-5688fbc24f48",
   },
   created_at: "2024-08-22T12:53:48.028702Z",
-}
+};
 
 const SESSION_DUMMY = {
-  access_token: 'test',
-  refresh_token: 'test',
+  access_token: "test",
+  refresh_token: "test",
   expires_in: 3,
-  token_type: 'test',
-  user:
+  token_type: "test",
+  user: USER_DUMMY,
+};
 
-}
+jest.mock("@/utils/supabase/component", () => ({
+  createClient: jest.fn(() => ({
+    auth: {
+      getUser: jest.fn(),
+      signInWithPassword: jest.fn(),
+      signUp: jest.fn(),
+    },
+  })),
+}));
 
+jest.mock("@/utils/storage/index", () => ({
+  userStorage: {
+    set: jest.fn(),
+    get: jest.fn(() => USER_DUMMY),
+  },
+}));
 
 describe("Supabase Authentication", () => {
   const supabase = createClient();
@@ -58,7 +53,7 @@ describe("Supabase Authentication", () => {
     it("유저 정보를 성공적으로 반환해야 한다", async () => {
       const getUserResponse: UserResponse = {
         data: {
-          user: USER_DUMMY
+          user: USER_DUMMY,
         },
         error: null,
       };
@@ -74,7 +69,9 @@ describe("Supabase Authentication", () => {
     });
 
     it("인증 서버 에러 발생 시 CustomError를 던져야 한다", async () => {
-      jest.spyOn(supabase.auth, "getUser").mockRejectedValue(new Error("Server Error"));
+      jest
+        .spyOn(supabase.auth, "getUser")
+        .mockRejectedValue(new Error("Server Error"));
 
       await expect(authenticateUser()).rejects.toThrow(CustomError);
     });
@@ -85,11 +82,13 @@ describe("Supabase Authentication", () => {
       const mockResponse = {
         data: {
           user: USER_DUMMY,
-          session: SESSION_DUMMY
+          session: SESSION_DUMMY,
         },
-        error: null
+        error: null,
       };
-      jest.spyOn(supabase.auth, "signInWithPassword").mockResolvedValue(mockResponse);
+      jest
+        .spyOn(supabase.auth, "signInWithPassword")
+        .mockResolvedValue(mockResponse);
 
       const result = await requestToSignIn("test@example.com", "password");
 
@@ -102,7 +101,9 @@ describe("Supabase Authentication", () => {
     });
 
     it("로그인 에러 시 CustomError를 던져야 한다", async () => {
-      jest.spyOn(supabase.auth, "signInWithPassword").mockRejectedValue(new CustomError("Login Error"));
+      jest
+        .spyOn(supabase.auth, "signInWithPassword")
+        .mockRejectedValue(new CustomError("Login Error"));
 
       await expect(
         requestToSignIn("test@example.com", "wrongpassword")
@@ -114,9 +115,9 @@ describe("Supabase Authentication", () => {
     it("회원가입 요청을 성공적으로 처리해야 한다", async () => {
       const mockResponse = {
         data: { user: USER_DUMMY, session: SESSION_DUMMY },
-        error: null
+        error: null,
       };
-      jest.spyOn(supabase.auth, 'signUp').mockResolvedValue(mockResponse);
+      jest.spyOn(supabase.auth, "signUp").mockResolvedValue(mockResponse);
 
       const result = await requestToSignUp("test@example.com", "password");
 
@@ -129,7 +130,9 @@ describe("Supabase Authentication", () => {
     });
 
     it("회원가입 에러 시 CustomError를 던져야 한다", async () => {
-      jest.spyOn(supabase.auth, "signUp").mockRejectedValue(new CustomError("SignUp Error"));
+      jest
+        .spyOn(supabase.auth, "signUp")
+        .mockRejectedValue(new CustomError("SignUp Error"));
 
       await expect(
         requestToSignUp("test@example.com", "password")
